requirements.txt
streamlit
google-generativeai
gTTS
Pillow
streamlit-camera